# Optimize LIKE with custom escape char