--- a/velox/benchmarks/basic/CMakeLists.txt
+++ b/velox/benchmarks/basic/CMakeLists.txt
@@ -63,9 +63,14 @@ add_executable(velox_benchmark_basic_preproc Preproc.cpp)
 target_link_libraries(velox_benchmark_basic_preproc ${velox_benchmark_deps}
                       velox_functions_prestosql velox_vector_test_lib)
 
-add_executable(velox_like_functions_benchmark LikeFunctionsBenchmark.cpp)
-target_link_libraries(velox_like_functions_benchmark ${velox_benchmark_deps}
+add_executable(velox_like_tpch_benchmark LikeTpchBenchmark.cpp)
+target_link_libraries(velox_like_tpch_benchmark ${velox_benchmark_deps}
                       velox_functions_lib velox_tpch_gen velox_vector_test_lib)
+
+add_executable(velox_like_benchmark LikeBenchmark.cpp)
+target_link_libraries(
+  velox_like_benchmark ${velox_benchmark_deps} velox_functions_lib
+  velox_functions_prestosql velox_vector_test_lib)
 
 add_executable(velox_benchmark_basic_vector_fuzzer VectorFuzzer.cpp)
 target_link_libraries(velox_benchmark_basic_vector_fuzzer