Currently we optimize LIKE operation only if escape char is not specified,
this PR adds the ability to apply the optimization even if user specifies
escape char. We introduced a PatternStringIterator which handles escaping 
transparently, so existing optimizations(kPrefix, kSuffix, kSubstring etc)
now work for patterns with escape char transparently, and future 
optimizations will have effect for escaped pattern transparently too. 

The benchmark result before this optimization:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like_generic##like_generic                                   4.48s   223.04m
----------------------------------------------------------------------------
----------------------------------------------------------------------------
like_prefix##like_prefix                                     1.26s   791.19m
like_prefix##starts_with                                  124.82ms      8.01
like_substring##like_substring                               4.23s   236.49m
like_substring##strpos                                    247.50ms      4.04
like_suffix##like_suffix                                     3.23s   309.58m
like_suffix##ends_with                                    127.47ms      7.85

```

After:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like_generic##like_generic                                   3.80s   263.33m
----------------------------------------------------------------------------
----------------------------------------------------------------------------
like_prefix##like_prefix                                  123.08ms      8.13
like_prefix##starts_with                                  123.66ms      8.09
like_substring##like_substring                            191.93ms      5.21
like_substring##strpos                                    237.96ms      4.20
like_suffix##like_suffix                                  127.67ms      7.83
like_suffix##ends_with                                    125.07ms      8.00
```

In Summary: 

- Speedup of kSubstring is about  20x.
- Speedup of kPrefix is about 10x.
- Speedup of kSuffix is about 20x.

We can confirm the speedup is reasonable from the comparison between our 
optmizations and the simple scalar function strpos, starts_with, ends_with, the 
performance numbers are quite close(see the like##strpos/starts_with/ends_with
in the benchmark result for more details).