@xumingming 2023-11-30
@mbasmanova Seems I have some misunderstanding here, I thought this PR would fix #7685 , for pattern `%xxx_xxx_xxx%`, I was thinking `_` here is a literal `_` rather than a single char wildcard, so both `_` and `%` here are wildcard here right?

But anyway, support LIKE optimization when the escape char is specified, right?