@mbasmanova 2023-11-30
@xumingming James, we are seeing users write queries like this: `SELECT * FROM t WHERE a LIKE b || '%' `.

Here, LIKE pattern appears as a prefix pattern, but when 'b' contains underscores the pattern is actually not a simple prefix patter. The user intent is of course for a prefix pattern, but we cannot assume that. The underscores happen often when a is a URL. Hence, it would be nice to optimize for prefix, suffix and substring patterns that contain underscores (wildchar).

> LIKE optimization when the escape char is specified is valuable, right?

Theoretically, yes, but I haven't seen these in practice.