@xumingming 2023-12-04
Add a benchmark to test the like optimization when escape char is specified, before this optimization:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##like_substring                                       22.59ms     44.26
like##like_prefix                                          14.08ms     71.00
like##like_suffix                                          20.18ms     49.56
like##like_generic                                         21.37ms     46.80
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```

after:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##like_substring                                      586.50us     1.71K
like##like_prefix                                         361.83us     2.76K
like##like_suffix                                         361.83us     2.76K
like##like_generic                                         21.40ms     46.72
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```