@xumingming 2023-12-05
Addressed all review comments, updated the benchmark to use longer strings, will paste the new benchmark result later.