@xumingming 2023-12-05
before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##like_substring                                       41.03ms     24.37
like##like_prefix                                          21.04ms     47.53
like##like_suffix                                          40.94ms     24.43
like##like_generic                                         40.86ms     24.47
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```

after:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##like_substring                                       78.73us    12.70K
like##like_prefix                                          42.70us    23.42K
like##like_suffix                                          42.76us    23.39K
like##like_generic                                         39.44ms     25.35
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```