@xumingming 2023-12-08
I’am not sure if there is something wrong with the design of the benchmark, the improvement seems huge.


Before(goes `Generic`):

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##substring                                            40.99ms     24.40
like##prefix                                               20.97ms     47.69
like##suffix                                               40.52ms     24.68
like##generic                                              40.41ms     24.74
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```

After(goes corresponding optimization):

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like##substring                                            83.71us    11.95K
like##prefix                                               44.34us    22.55K
like##suffix                                               42.37us    23.60K
like##generic                                              36.70ms     27.25
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```