@mbasmanova 2023-12-12
@xumingming James, have you been able to resolve the below concern?

> I’am not sure if there is something wrong with the design of the benchmark, the improvement seems huge.