@xumingming 2023-12-12
> @xumingming James, have you been able to resolve the below concern?
> 
> > I’am not sure if there is something wrong with the design of the benchmark, the improvement seems huge.

@mbasmanova After some experiments, I think there are two reasons for the huge performance improvement:

- Re2 is really slow compare to the optimizations we made, even if the input string is short(10 byte), Re2 is 100x slower than  our optimizations.
- When the input strings get longer(10bytes -> 1000bytes), the performance of our optimizations does not change much, but Re2's performance will be 10x slower.

And we can confirm the speedup is reasonable from the comparison between our optmizations and the simple scalar function strpos, starts_with, ends_with(as you suggested, really good idea :)), the performance nums are quite close(see the PR description for more details).