@xumingming 2023-12-12
> > I think there are two reasons for the huge performance improvement:
> 
> @xumingming Thank you for digging in and sharing your findings. These make sense. Let's add this context to the PR description to make it easily available for future readers. Awesome job!

Updated the PR description.